Fixing the saving of files related to Lategame Upgrades mod
The logic implemented before would always put a placeholder value in the list of LGU files whenever it detected a "LCSaveFile"
leading to cases like:

SaveFiles = { "LCSaveFile1" }
LGUFiles = { placeholder, "LGU_1" }

Ending up renaming the LGU file with a wrong identifier value and creating empty LGU files when entered into the same save (because it wasn't found a LGU file for that save)

The logic now is whenever we detect a save file, we check if there's any LGU file with the same identifier.
If so, add it to the lgu file list. Otherwise, add placeholder value instead.